Move ll.{c,h} to subdirectory, renaming ll.c to base.c
