+++ src/CMakeLists.txt
@@ -16,7 +16,7 @@ set(SOURCE_FILES
     libreset/avl.c
     libreset/ht.c
     libreset/bloom.c
-    libreset/ll.c
+    libreset/ll/base.c
     libreset/set.c
 )
