+++ src/libreset/avl.c
@@ -6,7 +6,7 @@
 #include "util/macros.h"
 #include "util/debug.h"
 
-#include "ll.h"
+#include "ll/ll.h"
 #include "avl.h"
 #include "util/macros.h"
