+++ src/libreset/avl.h
@@ -23,7 +23,7 @@
 #define __AVL_H__
 
 #include "bloom.h"
-#include "ll.h"
+#include "ll/ll.h"
 #include "util/attributes.h"
 #include "libreset/hash.h"
 #include "libreset/set.h"
