+++ src/libreset/ll/base.c
