+++ src/libreset/ll/ll.h
